refactor: Cutting out processing to each service
