+++ package.json
@@ -8,7 +8,7 @@
 		"dev": "node esbuild.config.mjs",
 		"build": "tsc -noEmit -skipLibCheck && node esbuild.config.mjs production",
 		"version": "node version-bump.mjs && git add manifest.json versions.json",
-		"test": "jest",
+		"test": "jest --config config/jest.config.ts",
 		"test-action": "jest --config config/jest.action.config.ts",
 		"test-watch": "jest --watchAll --config config/jest.config.ts"
 	},
