+++ src/types/type.ts
@@ -1,1 +1,2 @@
 export type StopPropagation = boolean;
+export const HEADINGS = [0, 1, 2, 3, 4, 5, 6];
